+++ ssh_server.go
@@ -109,9 +109,9 @@ func memoryCommand(op string, arg1 string, arg2 string) (res string) {
 				data[i+3] = byte(val & 0xff)
 			}
 		case "sd read":
-			data, err = usbarmory.SD.Read(int(addr), int(val))
+			data, err = usbarmory.SD.Read(int64(addr), int64(val))
 		case "mmc read":
-			data, err = usbarmory.MMC.Read(int(addr), int(val))
+			data, err = usbarmory.MMC.Read(int64(addr), int64(val))
 		}
 
 		res = hex.Dump(data)
